chore: refactor product placement section
